Crate protobuf_support

Supporting code for protobuf crates

Code in this crate is used in protobuf crates like protobuf or protobuf-parse. None of code in this crate has public API.

Modules

json_name

lexer

Implementation of lexer for both protobuf parser and for text format parser.

text_format

toposort